# forge-go/api/openapi/codegen-openapi.json
{
  "components": {
    "schemas": {
      "AddMessageToBoardRequest": {
        "properties": {
          "message_id": {
            "title": "Message Id",
            "type": "string"
          }
        },
        "required": [
          "message_id"
        ],
        "title": "AddMessageToBoardRequest",
        "type": "object"
      },
...
      "BoardMessagesResponse": {
        "properties": {
          "ids": {
            "items": {
              "type": "string"
            },
            "title": "Ids",
            "type": "array"
          }
        },
        "required": [
          "ids"
        ],
        "title": "BoardMessagesResponse",
        "type": "object"
      },
...
      "BoardsResponse": {
        "properties": {
          "boards": {
            "items": {
              "$ref": "#/components/schemas/BoardResponse"
            },
            "title": "Boards",
            "type": "array"
          }
        },
        "required": [
          "boards"
        ],
        "title": "BoardsResponse",
        "type": "object"
      },
...
      "CreateBoardRequest": {
        "properties": {
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "guild_id": {
            "title": "Guild Id",
            "type": "string"
          },
          "is_default": {
            "default": false,
            "nullable": true,
            "title": "Is Default",
            "type": "boolean"
          },
          "is_private": {
            "default": false,
            "nullable": true,
            "title": "Is Private",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "guild_id",
          "name",
          "created_by"
        ],
        "title": "CreateBoardRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IdInfo": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "IdInfo",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Rustic AI API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/nodes/register": {
      "post": {
        "operationId": "registerNode",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "node_id": {
                    "type": "string"
                  },
                  "capacity": {
                    "properties": {
                      "cpus": {
                        "type": "integer"
                      },
                      "memory": {
                        "type": "integer"
                      },
                      "gpus": {
                        "type": "integer"
                      }
                    },
                    "required": [
                      "cpus",
                      "memory"
                    ],
                    "type": "object"
                  }
                },
                "required": [
                  "node_id",
                  "capacity"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response"
          }
        },
        "summary": "Register Node"
      }
    },
    "/nodes/{node_id}/heartbeat": {
      "post": {
        "operationId": "nodeHeartbeat",
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          }
        },
        "summary": "Node Heartbeat"
      }
    },
...
    "/addons/boards/": {
      "get": {
        "description": "Retrieves all boards for a guild.\n\nArgs:\n    guild_id (str): The ID of the guild to get boards for.\n\nReturns:\n    BoardsResponse: The boards for the guild.",
        "operationId": "getBoards",
        "parameters": [
          {
            "in": "query",
            "name": "guild_id",
            "required": true,
            "schema": {
              "title": "Guild Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sqldb",
            "required": false,
            "schema": {
              "default": "sqlite:///rustic_app.db",
              "title": "Sqldb",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Boards",
        "tags": [
          "boards"
        ]
      },
      "post": {
        "description": "Creates a new board for a guild.\n\nArgs:\n    guild_id (str): The ID of the guild to create the board for.\n    board (CreateBoardRequest): The board creation request.\n\nReturns:\n    IdInfo: The id of the newly created board.",
        "operationId": "createBoard",
        "parameters": [
          {
            "in": "query",
            "name": "sqldb",
            "required": false,
            "schema": {
              "default": "sqlite:///rustic_app.db",
              "title": "Sqldb",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBoardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Board",
        "tags": [
          "boards"
        ]
      }
    },
    "/addons/boards/{board_id}/messages": {
      "get": {
        "description": "Retrieves all message IDs for a board.\n\nArgs:\n    board_id (str): The ID of the board to get messages for.\n\nReturns:\n    BoardMessagesResponse: The message IDs in the board.",
        "operationId": "getBoardMessageIds",
        "parameters": [
          {
            "in": "path",
            "name": "board_id",
            "required": true,
            "schema": {
              "title": "Board Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sqldb",
            "required": false,
            "schema": {
              "default": "sqlite:///rustic_app.db",
              "title": "Sqldb",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardMessagesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Board Messages",
        "tags": [
          "boards"
        ]
      },
      "post": {
        "description": "Adds a message to a board.\n\nArgs:\n    board_id (str): The ID of the board to add the message to.\n    request (AddMessageToBoardRequest): The request containing the message_id.\n\nReturns:\n    dict: Success message.",
        "operationId": "addMessageToBoard",
        "parameters": [
          {
            "in": "path",
            "name": "board_id",
            "required": true,
            "schema": {
              "title": "Board Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sqldb",
            "required": false,
            "schema": {
              "default": "sqlite:///rustic_app.db",
              "title": "Sqldb",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMessageToBoardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Message To Board",
        "tags": [
          "boards"
        ]
      }
    },
    "/addons/boards/{board_id}/messages/{message_id}": {
      "delete": {
        "description": "Removes a message from a board.\n\nArgs:\n    board_id (str): The ID of the board to remove the message from.\n    message_id (str): The ID of the message to remove.\n\nReturns:\n    dict: Success message.",
        "operationId": "removeMessageFromBoard",
        "parameters": [
          {
            "in": "path",
            "name": "board_id",
            "required": true,
            "schema": {
              "title": "Board Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "title": "Message Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sqldb",
            "required": false,
            "schema": {
              "default": "sqlite:///rustic_app.db",
              "title": "Sqldb",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Message From Board",
        "tags": [
          "boards"
        ]
      }
    },
    "/api/guilds": {
      "post": {
        "description": "Creates a new guild and adds it to the database.\n\nArgs:\n    launch_req (LaunchGuildReq):  object with spec for the new guild and the org id\n\nReturns:\n    IdInfo: The id of the newly created guild",
        "operationId": "createGuild",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LaunchGuildReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Guild"
      }
    },
    "/api/guilds/{guild_id}": {
      "get": {
        "description": "Retrieves details for a guild based on the provided guild_id.\n\nArgs:\n    guild_id (str): The ID of the guild to retrieve details for.\n\nReturns:\n    GuildSpecResponse: The retrieved guild details, or None if the guild was not found.\n\nRaises:\n    HTTPException: If the guild is not found, raises an HTTPException with a status code of 404.",
        "operationId": "getGuildDetailsById",
        "parameters": [
          {
            "in": "path",
            "name": "guild_id",
            "required": true,
            "schema": {
              "title": "Guild Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sqldb",
            "required": false,
            "schema": {
              "default": "sqlite:///rustic_app.db",
              "title": "Sqldb",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuildSpecResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Guild"
      }
    },
    "/api/guilds/{guild_id}/agents/{agent_id}/files/": {
      "get": {
        "operationId": "listFilesForAgent",
        "parameters": [
          {
            "in": "path",
            "name": "guild_id",
            "required": true,
            "schema": {
              "title": "Guild Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sqldb",
            "required": false,
            "schema": {
              "default": "sqlite:///rustic_app.db",
              "title": "Sqldb",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Files For Agent"
      },
      "post": {
        "operationId": "uploadFileForAgent",
        "parameters": [
          {
            "in": "path",
            "name": "guild_id",
            "required": true,
            "schema": {
              "title": "Guild Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sqldb",
            "required": false,
            "schema": {
              "default": "sqlite:///rustic_app.db",
              "title": "Sqldb",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uploadFileForAgent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MediaLink"
                  },
                  "title": "Response Uploadfileforagent",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload File For Agent"
      }
    },
    "/api/guilds/{guild_id}/agents/{agent_id}/files/{filename}": {
      "delete": {
        "operationId": "deleteAgentFile",
        "parameters": [
          {
            "in": "path",
            "name": "guild_id",
            "required": true,
            "schema": {
              "title": "Guild Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sqldb",
            "required": false,
            "schema": {
              "default": "sqlite:///rustic_app.db",
              "title": "Sqldb",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Files For Agent"
      },
      "get": {
        "operationId": "getFileForAgent",
        "parameters": [
          {
            "in": "path",
            "name": "guild_id",
            "required": true,
            "schema": {
              "title": "Guild Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "download",
            "required": false,
            "schema": {
              "default": false,
              "nullable": true,
              "title": "Download",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "sqldb",
            "required": false,
            "schema": {
              "default": "sqlite:///rustic_app.db",
              "title": "Sqldb",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get File For Agent"
      }
    },
    "/api/guilds/{guild_id}/files/": {
      "get": {
        "operationId": "listFilesForGuild",
        "parameters": [
          {
            "in": "path",
            "name": "guild_id",
            "required": true,
            "schema": {
              "title": "Guild Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sqldb",
            "required": false,
            "schema": {
              "default": "sqlite:///rustic_app.db",
              "title": "Sqldb",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Files For Guild"
      },
      "post": {
        "operationId": "uploadFileForGuild",
        "parameters": [
          {
            "in": "path",
            "name": "guild_id",
            "required": true,
            "schema": {
              "title": "Guild Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sqldb",
            "required": false,
            "schema": {
              "default": "sqlite:///rustic_app.db",
              "title": "Sqldb",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uploadFileForGuild"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload File For Guild"
      }
    },
...
  }
}